requirements.txt
numpy
scipy
pandas
sklearn
sktime
matplotlib
tqdm
rich
pytorch-lightning
hydra-core
munch
omegaconf
wandb
einops
opt_einsum
pykeops
cmake
torchtext
datasets
transformers
pytorch-fast-transformers
